Replace OFFSET_IGNITION with IGNITION_ADVANCE_LARGE.toUser/toRaw

+++ speeduino/units.h
@@ -129,7 +129,7 @@ static constexpr conversionFactor<int16_t, uint8_t> TPS_DOT = { .scale=10U, .tra
 static constexpr conversionFactor<uint16_t, uint8_t> CRANKING_ENRICHMENT = { .scale=5U, .translate=0U };
 
 /** @brief Ignition values from the main spark table are offset 40 degrees downwards to allow for negative spark timing */
-static constexpr conversionFactor<int8_t, uint8_t> IGNITION_ADVANCE_LARGE = { .scale=1U, .translate=-40 };
+static constexpr conversionFactor<int16_t, uint8_t> IGNITION_ADVANCE_LARGE = { .scale=1U, .translate=-40 };
 
 /** @brief Ignition advance adjustments can use a smaller offset */
 static constexpr conversionFactor<int8_t, uint8_t> IGNITION_ADVANCE_SMALL = { .scale=1U, .translate=-15 };
